Update replace.cue
updating the example for the capture group name in the example to use $$num instead of $num as single $ does not work. 
Also need to provide more description on the Note.
"Note that $foo is interpreted in a Vector configuration file, instead use $$foo." This isn't clear, and not sure what it means.