chore(observability): Simplify the default log targets selection
When Vector sets a log level, either by the default `info` level or by adding
`-v` or `-q` command-line options, it applies that log level to the following
targets: `vector`, `codec`, `vrl`, `file_source`, `tower_limit`, `rdkafka`,
`buffers`, `lapin`, and `kube`. These targets were set up to limit one or more
of the _unlisted_ targets to reduce very chatty debug logs. By contrast, if the
user specifies the environment variable setting `LOG_LEVEL=debug`, none of those
targets are used.  This change drops all those hard-coded targets and sets all
targets to the same level. If we discover there is still an issue with a chatty
target, we can add an override for just that target instead of this list.