chore: Shorten name of `skip_serializing_if_default` (#19591)
The actual code in the function merely compares the value to its type default,
so use a shorter name that reflects its actual function. This results in a serde
condition of `skip_serializing_if = "is_default"` which is a more natural
reading.  This ends up allowing `cargo fmt` to wrap more lines together,
dropping 100 LOC.