chore(datadog_logs sink): allow batching up to full payload limit
Closes #9202

Since #19189 was merged, this heuristic to try to avoid oversized requests is
no longer necessary from a correctness point of view. The only potential reason
to keep it would be if we expected oversized batches to be common, which could
mean a performance impact if the new batch-splitting code is triggered more
often to avoid oversized requests.

Another option would be to simply reduce buffer we leave ourselves between the
goal and the max, but any analysis of the best value would be entirely
dependent on the format of the event data.

Signed-off-by: Luke Steensen <luke.steensen@gmail.com>