fix(http_server source): Conditionally send Connection: Close header based on HTTP version
Previously this header was sent when `default_max_connection_age` expired for all HTTP requests
regardless of HTTP version, but this header is not applicable to HTTP/2 and HTTP/3 requests. Instead
only send this header for other HTTP versions.

HTTP/2 and HTTP/3 support GOAWAY frames for this but I think this may need to be implemented at
a different level.

Fixes: #19643

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>