chore(dev): Update CONTRIBUTING.md docs regarding how to have website preview created
To reduce the number of branch deploys since we have a hard cap on them in Amplify, restrict preview
builds to only branches with `website` in the name. Because Amplify automatically picks up created
branches, the only mechanism we seem to have to be able to filter them out at the moment is the
name.

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>