test(dnsmsg_parser): fix tests for currently unknown rdata types (#20052)
This changes tests for RData type currently unknown to `hickory_proto` to
use a common mechanism that is used for known types too. The goal of this
is to use a similar code path to an actual use case (which will first
consider known types and then try to treat them as unknown).

In #19921 implementation for HINFO was added and there was a test that
has treated HINFO as an unknown type that was run successfully, but in
real application use case, it would fail for such records, since it would
never get treated as an unknown type. This will help spot if any new types
,that are already supported by vector, have been added to `hickory_proto`.

Related: #19921