chore: DRY `into_variant()` via `akin` crate
This should be simple and lightweight enough to justify for the DRY benefit?

This way the method doesn't need to be duplicated redundantly. That was required because there is no trait for `FromRepr` provided via `strum`. That would require a similar amount of lines for the small duplication here.

The `akin` macro duplicates the `impl` block for each value in the `&enums` array.