chore(deps): Bump MSRV to reduce usage of `async_trait`
There are several uses that still require the use of `async_trait`, namely those
where the resulting trait must be object safe (i.e. is used in a `Box<dyn
Future>` construct). The `async_trait` macro handles this by boxing the
resulting future, so we _could_ work around this ourselves, but only by manually
expanding the boxing ourselves, which is not a win over just using the trait.