fix: Support deserializing config value that is a number type
In a YAML config a string can optionally be wrapped with quotes, while a number that isn't quote wrapped will be treated as a number type.

The current support was only for string numbers, this change now supports flexibility for config using ordinal values in YAML regardless of quote usage.

The previous `Self::into_variant(&s)` logic could have been used instead of bringing in `serde-aux`, but the external helper attribute approach seems easier to grok/follow as the intermediary container still seems required for a terse implementation.

The match statement uses a reference (_which requires a deref for `from_repr`_) to appease the borrow checker for the later borrow needed by `value` in the error message.