docs: fix grammar in what-is-observability-pipelines.md (#20387)
fix grammar in what-is-observability-pipelines.md

I didn't change the name of the file (although that should really be changed to `what-are-observability-pipelines.md`) because I didn't want to break links.  However, if you're willing to create a forwarding rule in your web server, it would be ideal if the file name and URL had `are`.