chore(core): Allow arbitrary data types in `registered_event!`
The data type parameter for the `emit` event function in the `registered_event!`
macro is specified as an `ident`. This restricts the possible values to named
types like tuples and newtype wrappers, which is the desired pattern, but that
presents problems for passing in references. This change loosens the type in the
macro to `ty` which allows for arbitrary types to be used.