fix(datadog sinks): Compute proper validate endpoint (#20644)
* fix(datadog sinks): Compute proper validate endpoint

When no `endpoint` is configured in the `datadog_*` sinks, they use the `site`
instead to compute separate endpoints for the API key validation and for the
actual service. However, when an `endpoint` _is_ configured, that URI is used
for both, leading to a HTTP 404 Not Found error when trying to run the health
check.

This change pulls in the API endpoint computation used by the Datadog Agent to
rewrite the `endpoint` URI when it is detected as a valid Datadog domain.

* Add spelling allow for "ddog"

* Fix changelog grammar