fix(kafka source): consumer subscribe in main kafka source task
This moves the consumer.subscribe(topics) call into the main kafka
source loop, to ensure that callback handlers will be set up whenever
the consumer is subscribed.

In `vector validate` mode for example, the consumer sometimes receives
partition assignments, and the consumer invokes pre_rebalance handlers.
The main consumer task is not polled in that scenario, so callback
messages sent to the coordination task are never handled and the
pre_rebalance handler deadlocks. Moving the subscribe call into the main
loop means that the consumer will not join the group and will not get
a partition assignment unless the main source task is being polled.