chore(ci): Have Vector `master` use `main` of VRL
So that nightly releases can you unreleased VRL changes for testing. This was requested by a user.
I can see the usefulness to be able to validate VRL changes without having to build Vector.

As part of the release, we'll switch it to the latest released version (which should be the same
SHA).

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>