enhancement(aws provider): Add ability to disable request signing
This can be useful when sending anonymous requests to AWS S3. Potentially it could be useful in
other situations as well (e.g. sending to AWS API compatible endpoints that don't support signing).

I'd really have liked to introduce this as a new "strategy" for AWS authentication configuration
since it is mutually exclusive with the others, but given the way AWS authentication configuration
is currently implemented as an untagged enum, adding it to the default config enum seemed like the
best option.

If/when we refactor this to follow
https://github.com/vectordotdev/vector/blob/master/docs/specs/configuration.md#polymorphism then we
can move it.

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>