<fix>: Correct snafu features (#21007)
Enable std feature on snafu

Snafu does not build if the futures feature is enabled and the
std feature is disabled. Snafu was already being built with
the std feature enabled because some other dependency of
vector happened to enable it. This commit just makes the
existing implicit feature dependency explicit instead.