enhancement(config): Remove the watcher's coupling to SIGHUP / prepare for automatic Windows config reload (#20989)
* Remove the watcher's dependency on SIGHUP / unix

Right now the watcher is broken/unimplemented for Windows and a large part of the
reason for that might be how it works:

* the watcher subscribes to filesystem events (crate: notify)
* the watcher raises a unix signal (SIGHUP) on detecting changes
* the signal handler (again, unix only) reacts to SIGHUP with an internal
  reload signal
* (vector magically reloads/compares the configs)

With these changes that unix signal dependency is removed

* the watcher clones a SignalTx to emit its very own signals
* the watcher sends an internal reload signal on detecting changes
* (vector magically reloads/compares the configs)

All of this _should_ also work on Windows, has nothing to do with unix signals anymore.

I slightly modified the tests to not just wait for any Ok, but to state explicitly what
signal we expect to receive.

Cargo fmt

* Make clippy happy

* Add changelog entry

* Update changelog.d/support-config-watch-for-windows.enhancement.md

* Fix newly merged test case

---------

Co-authored-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>