chore(config): Configure `datadog_search` condition directly
The `datadog_search` condition previously was configured with a generic string,
which would accept any text. When that condition was built, it would then parse
that search string into a query data structure. Since that query data structure
now supports deserialization directly, we can use that as the configuration type
and skip the extra parse step.

This should not change any user facing behavior except for the exact wording of
the error message for invalid query strings.