enhancement(core): Wrap event metadata in `Arc` for clone performance
The `EventMetadata` structure contains a number of sub-structures that can cause
extensive deep copies when an event is cloned, most notably `value:
vrl::value::Value`. To improve this, wrap the actual inner event metadata in an
`Arc`, and use the standard tools from that type to only clone it when needed to
provide a mutable or owned copy.