Allow reading the git short has from the build environment
Why are we doing this whole handling of the environment variable? Packagers often do not package
from a git working copy? they grab a tarball and `cargo vendor` the dependencies and then build in
an offline environment. This allows them to give the value of the hash (or an equivalent value like
the tag) to the build from the build description. If that value is not in the environment, we fall
through to the old code that parses the git rev-parse output.

So from all the packagers: Thank you for keeping this code alive!