fix(sources): fix unexpected timestamp field
When log namespacing is enabled, `exec` and `http_server` sources include
a redundant `timestamp` field when decoding is enabled. When decoding
is not enabled, they put the content under a _"message"_ key, instead
of directly in the root object.

In the case of `exec`, the issue was caused by passing a hardcoded
`LogNamespace` value to the decoder. Fixed by passing the parsed value.

In the case of `http_server`, the decoder ignored the global
`LogNamespace`. Fixed by passing the correctly parsed value to the
decoder.

Fixes #20404.