chore(dev): Add configuration for aqua to manage tooling dependencies (#18020)
* chore(dev): Add configuration for aqua to manage tooling dependencies

Aqua is a tool dependency manager: https://aquaproj.github.io/

Here I introduce configuration just for the Rust-based tooling we use.

My intent is to try it out for a bit (along with anyone else who is interested) before trying to
integrate it more tightly into the development processes (e.g. mentioning in DEVELOPING.md and using
in CI).

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

* spelling

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

* Add dd-rust-license-tool too

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

---------

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>