docs(vrl): document punycode as an IDN function (#21734)
* docs(vrl): document punycode as an IDN function

This adds a note to `encode_punycode` and `decode_punycode` to inform users that it is meant to be
used in IDN context and that passed values are handled like domain names. This is important because
punycode by default doesn't use `xn--` prefix, but it is added in IDN context.

Related: https://github.com/vectordotdev/vrl/pull/1104

* Update wording in decode_punycode docs

Co-authored-by: Alicia Scott <aliciascott@users.noreply.github.com>

---------

Co-authored-by: Alicia Scott <aliciascott@users.noreply.github.com>