Add `SinkConfig` impl for `MemoryConfig`
This adds a `SinkConfig` implementation for `MemoryConfig`, making it a shared configuration for
both enrichment table component and the sink component. To ensure that the exact same `Memory`
instance is used, created `Memory` is cached inside the config instance and then shared whenever one
of the `build` variants are invoked.

`Memory` is already built with sharing between different instances in mind, since it had to be
shared across different threads and it is cloneable while keeping the same backing structure for
data.

The actual sink building is a placeholder.