feat(internal): Add VECTOR_HOSTNAME env variable (#21789)
* Add VECTOR_HOSTNAME env variable

By default, Vector uses the hostname returned by the gethostname. However, in cases when Vector is running in containers, it is meaningless because hostname is always the name of the container, e.g., the name of a pod in Kubernetes cluster.

This commit adds the way to override a hostname by using the env variable. In Kubernetes it can be used like this to see the name of a node where the Vector process is running (which is more meaningful).

```
env:
- name: VECTOR_HOSTNAME
  valueFrom:
    fieldRef:
      fieldPath: spec.nodeName
```

Signed-off-by: maksim.nabokikh <max.nabokih@gmail.com>

* Add changelog and docs, fix the naming.

Signed-off-by: maksim.nabokikh <max.nabokih@gmail.com>

* Update changelog.d/21789_add_vector_hostname_env_variable.enhancement.md

Co-authored-by: Ursula Chen <58821586+urseberry@users.noreply.github.com>

* - Use variables from runtime
- Fix cue formatting
- Use the variable value if set but empty

Signed-off-by: maksim.nabokikh <max.nabokih@gmail.com>

---------

Signed-off-by: maksim.nabokikh <max.nabokih@gmail.com>
Co-authored-by: Ursula Chen <58821586+urseberry@users.noreply.github.com>