fix(config): correctly propagate http config provider Result to caller (#21982)
Problem: The HTTP config provider has been failing to parse the provided
TOML from its endpoint when polled. My best guess is that including the
`?` on the end of the http provider load results in the TOML parser
subtly exploding due to the aliased return type (`BuildResult`).

It appears that this was broken in
`8ed9ec24c1eba0b2191d7c1f24ec2a7540b5bebf`.

See https://github.com/vectordotdev/vector/issues/21873 for the initial
issue report.

Solution: Correctly propagate the `ConfigBuilder` by not using the `?`
operator. This appears to allow the TOML parser to correctly pick up on
the return type.

-----

Thanks to @jszwedko for the bisect, saved a ton of pain for me.

closes https://github.com/vectordotdev/vector/issues/21873