feat!(gcp_stackdriver_logs sink): Support more fields
The LogEntry[1] type supports a lot more fields than are currently
supported by the sink. This change adds support for labels, but more
fields can be added the same way.

Additionally this removes the timestamp from the message, if set,
instead of having it in both the 'timestamp' field, as well as part of
the message payload.

It also changes the payload type from json to text if the message is
only a string. This makes the GCP search queries simpler, e.g.
`textPayload: "some message" instead of `jsonPayload.msg: "some
message".

[1]: https://cloud.google.com/logging/docs/reference/v2/rest/v2/LogEntry