fix(gcp service): Prevent overflow for gcp token expiry calculation (#22639)
- The simple calculation for the next deadline to sleep the gcp refresh
token thread performs a subtraction which is unchecked.

- The value of `expires_in` may be less then the constant
METADATA_TOKEN_EXPIRY_MARGIN_SECS, causing the deadline to overflow and
making the thread sleep for a long duration. Upon the next time the sink
pushes to GCS the token would have already expired since the refresh
thread hadn't awaken on time.