Copy GCP stackdriver logsink into opentelemetry
The HTTP sink opentelemetry is based on currently doesn't seem to be
extensible enough for what we need.

Looking at other sinks they all seem to be based on just the things they
need, so this takes the same approach. The GCP sink seems like a good
starting point for this.