source opentelemetry: Add support for HTTP/JSON
This adds support for decoding the otlp payload as json. Making it
possible to send json instead of having to add protobuf support to
applications.

The implementation is a bit involved as both the parsing and the
reply depend on the content type used, so they must propagate
properly.
This is done by extracting it from the incoming request and passing it
along all the way to the response handler.

In the process it also fixes how errors are handled. They're currently
propagated using `warp::reject::custom`. However `reject` means that the
request is passed on to another handler. This behaviour isn't desired,
we only have one handler per path and there won't be another match.
Instead we must send the error out during handling, which will also
give us access to the content type.
The status code must also be set when the request failed to deserialize