Fix for data loss in datadog logs sink w/ logs namespacing enabled
- When logs namespacing is enabled on the datadog logs sink data loss will be
observed. This is due to a bug that was introduced in a method called normalize_event
 which makes events conform with a standard that the datadogs logs backend defines.

- The cause is due to unexpected behavior in how some of the methods in LogEvent.rs behave
when the underlying Value type is not an Object. When it is not an object the value will
be coerced into one and the existing data that was in the value type would be lost. This
is why some of the existing unit tests passed, because in those tests the input was hardcoded
to the type of an object, whereas coming from the datadog_agent source, the Event was of type
Bytes.

- The fix is to ahead of time coerce the type into an object (if necesssary) and nest it under
the 'message' key, where the datadog logs backend expects the content of the log to exist.