vdev: do not terminate before child on SIGINT
The SIGINT(control-c) is sent to both the parent and child, as they're
in the same group. The `vdev` process quits immediately, as its just
waiting on the child. However the child will go through its shutdown
process, which by default takes a minute if it cannot cleanly shutdown.
During this minute your controlling terminal gets flooded with shutdown
messages.

The way to stop it is to `kill` vector again, e.g. `pkill vector` but
its annoying and could kill other vector instances.

This lets the vdev ignore the SIGINT and just rely on the parent to
terminate after the child did.

Its currently only implemented on unix using the libc crate. Vector uses
the nix crate, but that seemed a bit overkill for 2 lines.