add content-type extractor
This will allow us to support json in a later stage, as the content type
of the reply must match what the client used in the request.

This also improves error messages a bit, as its now a 415 instead of a
500. A 500 internal server error is quite confusing when debugging.
There is no error with the server, its the client sending the wrong
request.