migrate from warp::rejection to reply
Reject and recover indicates that the current filter didn't match but
that others might[1]. This means reject is not the appropriate tool for
handling e.g. deserialization errors.

This commit migrates from rejection to returning the reply directly, and
at the same time updates it to the new status codes to make sure it
passes the new tests.

This change also makes it possible to support `application/json` in the
future. The `content-type` of the reply must match that of the request,
and as we now have access to the request headers we can make sure they
match.

[1]: https://github.com/seanmonstar/warp/issues/388#issuecomment-576453485