fix(datadog_logs sink): Remove `_collisions` field in agent normalization routine (#22956)
* fix(datadog_logs sink): Remove _collisions field in agent normalization

- In previous commits to add the ability to mimic the agent payload,
logic was added to take any fields that would collide and store them in
a new object at the root called '_collisions'

- Without this option enabled if users attempt to assemble a payload
with a key at the object root, that matches a key nested under 'message'
then the logs-intake when unraveling the message, will note a duplicate
item.

- This normalization routine should follow the same behavior and
therefore nest collisions under a nested 'message' key which would in
turn keep the behavior of the system consistent and have the added
advantage of not polluting the users payloads with additional fields.

* Removing non-relevant unit test

- In a previous implementation the normalize_as_agent method would
deserialize the payload contents depending on its type.

- This is no longer performed therefore this unit test is considered
moot.