feat(remap transform): make `--watch-config` watch external VRL files in `remap` transforms (#23010)
* feat(transforms): add files_to_watch support for transforms

Invoke the `files_to_watch` function for applicable transforms to
collect external file paths that should be watched for changes.

* feat(remap transfrom): add files_to_watch to track VRL files

Add a `files_to_watch` function to the remap transform to collect file
paths from the `file` and `files` fields. This allows to watch these
files for changes and trigger reloads when the VRL code is modified.

* fix(config): relative path handling for external files when `--watch-config` enabled

Relative paths to external files in the configuration are now correctly
resolved when using --watch-config.
Previously, changes to such files not trigger a reload if the path
provided in the config was relative. This fix ensures that relative
paths are canonicalized, allowing file changes to be properly detected.