enhancement(splunk_hec sink): Use a response cookie to route ack checks to same indexer
Closes #19417.

This is particularly useful when running Splunk in a clustered
environment with multiple indexer hosts. In this environment,
acknowledgement IDs are frequently duplicated across multiple indexers
(they all start at `0` and count upwards as they receive requests with
the same `X-Splunk-Request-Channel` header, so there will be lots of
reuse). One common way to distinguish between multiple hosts behind a
load balancer is to return a cookie to specify which indexer to respond
back to. This is the recommended way, for instance, to set up an AWS ELB
for a Splunk indexer cluster such that it has cookie stickiness enabled:
https://docs.splunk.com/Documentation/AddOns/released/Firehose/ConfigureanELB
https://community.splunk.com/t5/Getting-Data-In/How-to-configure-the-load-balancer-to-handle-HEC-data/td-p/742116