fix(sources): Handle sources with no default output (#23172)
* fix(sources): Handle sources with no default output

There exist sources that can be configured to output events entirely to
non-default ports. The refactoring to `struct SourceSender` in #23089 failed to
take this into account. This PR reverts the changes to the `default_output`
handling in the builder and `SourceSender` with the addition of a small cleanup
of its own.

The sources that can have this condition are `datadog_agent` when
`multiple_outputs` is set and `opentelemetry`.

* Generalize comment