chore(config): Fix `InlineSingleUseReferencesVisitor` failing tests
The above visitor is intended to inline schemas in the references section for
which there is only a single reference. This visitor is, however, unable to
detect some cases where there are in fact multiple references, ending up
inlining the schema multiple times. This has led to bloat in generated schemas.

In the core Vector schema, this affects 121 structures, shaving 844 lines off of
the generated schema.