docs(external): Updated `replace` capture group example  (#18785)
* Update replace.cue

updating the example for the capture group name in the example to use $$num instead of $num as single $ does not work. 
Also need to provide more description on the Note.
"Note that $foo is interpreted in a Vector configuration file, instead use $$foo." This isn't clear, and not sure what it means.

* Update replace.cue

Added additional example for Replace with capture group to give an example of calling capture groups in a configuration file and outside a configuration file so it is clear.

* Update website/cue/reference/remap/functions/replace.cue

* Update website/cue/reference/remap/functions/replace.cue

Co-authored-by: May Lee <mayl@alumni.cmu.edu>

* Update website/cue/reference/remap/functions/replace.cue

Co-authored-by: May Lee <mayl@alumni.cmu.edu>

* test case fix

* cue fmt

---------

Co-authored-by: Jesse Szwedko <jesse@szwedko.me>
Co-authored-by: May Lee <mayl@alumni.cmu.edu>
Co-authored-by: Pavlos Rontidis <pavlos.rontidis@gmail.com>