enhancement(redis sink): support stream data structure in redis sink
Redis has support for a newish data structure called streams that model unbounded data streams better than lists or PUBSUB do. Consumers can use a cursor to move through a stream reliably with redis assigning ids to individual records in the stream. Streams also support automatic trimming to keep the stored length bounded, and clients can implement Kafka-style consumer groups to process a stream in parallel. This adds support for publishing to redis streams using the XADD command, and an optional config to tell Redis to keep the length of the stream within a max.