Drop clamping u64 values to "safe" floats
The "safe" conversion from u64 values into f64 clamped integers that cannot be
exactly represented to the maximum safe integer value. These converted values
were always used to `set` a gauge value. When looking at that gauge, however,
retaining the magnitude is more important than keeping it exactly representable
as an integer, so the safe conversion is actually lossy in the wrong direction.