chore(transforms): Expose Datadog search matching directly
Searching using the `DatadogSearchRunner` is exposed through the `Conditional`
trait. However, due to other use cases (namely VRL), that trait takes ownership
of the event and returns a new copy of it. This ownership needlessly complicates
the calling convention if we just want to use the search running, as the
underlying VRL method only needs to borrow the event. This PR exposes the search
interface directly, borrowing the event.