enhancement(datadog_agent source): Add control for metric name splitting (#23986)
In #13176, the `datadog_agent` source had functionality introduced to
automatically split the metric series name into namespace and name components on
the first period in that name.  This change was done to address #12953 where a
default namespace was configured being applied to a metrics sink, resulting in
metrics from this source having an undesired prefix added.

Since all of the metric sinks will reproduce the metric series name based on
both the namespace and name in the same way, this is a lossless transform.
However, if we want to filter on the full name (ex in a remap VRL script etc),
having the namespace split off becomes a problem.

To address this problem, this change introduces a `split_metric_namespace`
option to the `datadog_agent` source. It defaults to `true`, which is the
current behavior, so the change is entirely backward compatible with existing
configs.

The bulk of the size of the change comes from the addition of new tests and
refactoring existing tests to reduce duplicated code.