fix(enrichment tables): prevent crash on config reload with enrichment table sources
Diff calculation considered source keys (which are different than component keys in the case of
enrichment tables) to find differences in configuration. This caused crashes when that component
couldn't be found, because it has a different key in the configuration. This fix changes that to use
the original component key for such lookups.