fix(enrichment tables): prevent crash on config reload with enrichment table sources (#24014)
* fix(enrichment tables): prevent crash on config reload with enrichment table sources

Diff calculation considered source keys (which are different than component keys in the case of
enrichment tables) to find differences in configuration. This caused crashes when that component
couldn't be found, because it has a different key in the configuration. This fix changes that to use
the original component key for such lookups.

* Add changelog entry

* Add unit test for the fix

* Change tests to yaml config

* improve readability

* improve readability

---------

Co-authored-by: Pavlos Rontidis <pavlos.rontidis@gmail.com>